Split background worker into modules
